refactor: unify test cometbft block construction
