+++ crates/astria-conductor/Cargo.toml
@@ -49,9 +49,9 @@ telemetry = { package = "astria-telemetry", path = "../astria-telemetry" }
 [dev-dependencies]
 jsonrpsee = { workspace = true, features = ["server"] }
 
-sequencer-types = { package = "astria-sequencer-types", path = "../astria-sequencer-types", features = [
-  "test-utils",
-] }
+# sequencer-types = { package = "astria-sequencer-types", path = "../astria-sequencer-types", features = [
+#   "test-utils",
+# ] }
 optimism = { package = "astria-optimism", path = "../astria-optimism", features = [
   "test-utils",
 ] }
@@ -59,6 +59,6 @@ config = { package = "astria-config", path = "../astria-config", features = [
   "tests",
 ] }
 proto = { package = "astria-proto", path = "../astria-proto", features = [
-  "server",
+  "server", "test-utils",
 ] }
 tokio-stream = { version = "0.1.14", features = ["net"] }
