+++ crates/astria-conductor/src/data_availability/block_verifier.rs
@@ -281,6 +281,7 @@ mod test {
     };
 
     use proto::native::sequencer::v1alpha1::{
+        test_utils::make_cometbft_block,
         RollupId,
         UncheckedCelestiaSequencerBlob,
     };
@@ -364,7 +365,7 @@ mod test {
         let rollup_transactions_proof = tree.construct_proof(0).unwrap();
         let rollup_ids_proof = tree.construct_proof(1).unwrap();
 
-        let mut header = sequencer_types::test_utils::default_header();
+        let mut header = make_cometbft_block().header;
         let height = header.height.value().try_into().unwrap();
         header.data_hash = Some(Hash::try_from(data_hash.to_vec()).unwrap());
 
@@ -402,7 +403,7 @@ mod test {
         let rollup_transactions_proof = tree.construct_proof(0).unwrap();
         let rollup_ids_proof = tree.construct_proof(1).unwrap();
 
-        let mut header = sequencer_types::test_utils::default_header();
+        let mut header = make_cometbft_block().header;
         let height = header.height.value().try_into().unwrap();
         header.data_hash = Some(Hash::try_from(data_hash.to_vec()).unwrap());
