+++ crates/astria-conductor/src/sequencer/mod.rs
@@ -402,7 +402,7 @@ mod tests {
         },
         stream::FuturesOrdered,
     };
-    use sequencer_types::test_utils::create_tendermint_block;
+    use proto::native::sequencer::v1alpha1::test_utils::make_cometbft_block;
 
     use super::{
         forward_block_or_resync,
@@ -440,9 +440,9 @@ mod tests {
     #[tokio::test]
     async fn block_at_expected_height_is_forwarded() {
         let expected_height = 5u32.into();
-        let mut tendermint_block = create_tendermint_block();
-        tendermint_block.header.height = expected_height;
-        let expected_block = SequencerBlock::try_from_cometbft(tendermint_block)
+        let mut block = make_cometbft_block();
+        block.header.height = expected_height;
+        let expected_block = SequencerBlock::try_from_cometbft(block)
             .expect("the tendermint block should be well formed");
 
         let mut env = ForwardBlockOrResyncEnvironment::setup().await;
@@ -481,9 +481,9 @@ mod tests {
     async fn future_block_triggers_resync() {
         let expected_height = 5u32.into();
         let future_height = 8u32.into();
-        let mut tendermint_block = create_tendermint_block();
-        tendermint_block.header.height = future_height;
-        let expected_block = SequencerBlock::try_from_cometbft(tendermint_block)
+        let mut block = make_cometbft_block();
+        block.header.height = future_height;
+        let expected_block = SequencerBlock::try_from_cometbft(block)
             .expect("the tendermint block should be well formed");
 
         let mut env = ForwardBlockOrResyncEnvironment::setup().await;
@@ -510,9 +510,9 @@ mod tests {
     #[tokio::test]
     async fn older_block_is_dropped() {
         let expected_height = 5u32.into();
-        let mut tendermint_block = create_tendermint_block();
-        tendermint_block.header.height = 3u32.into();
-        let expected_block = SequencerBlock::try_from_cometbft(tendermint_block)
+        let mut block = make_cometbft_block();
+        block.header.height = 3u32.into();
+        let expected_block = SequencerBlock::try_from_cometbft(block)
             .expect("the tendermint block should be well formed");
 
         let mut env = ForwardBlockOrResyncEnvironment::setup().await;
