+++ crates/astria-proto/src/native/sequencer/v1alpha1/mod.rs
@@ -25,6 +25,9 @@ use crate::{
     },
 };
 
+#[cfg(feature = "test-utils")]
+pub mod test_utils;
+
 pub const ADDRESS_LEN: usize = 20;
 pub const ROLLUP_ID_LEN: usize = 32;
