+++ crates/astria-sequencer-relayer/Cargo.toml
@@ -46,6 +46,7 @@ config = { package = "astria-config", path = "../astria-config", features = [
   "tests",
 ] }
 merkle = { package = "astria-merkle", path = "../astria-merkle" }
+proto = { package = "astria-proto", path = "../astria-proto", features = ["test-utils"] }
 sequencer-types = { package = "astria-sequencer-types", path = "../astria-sequencer-types", features = [
   "test-utils",
 ] }
