+++ crates/astria-sequencer-types/src/lib.rs
@@ -2,7 +2,4 @@ pub mod abci_code;
 pub mod cometbft;
 pub mod serde;
 
-#[cfg(feature = "test-utils")]
-pub mod test_utils;
-
 pub use abci_code::AbciCode;
